Add INCR command
Storage::incr atomically parses the current value as an integer
(defaulting missing keys to 0), increments it, and persists the
result via the existing AOF SET format so it replays correctly on
reload and replicates to replicas like a normal write. Non-integer
values return the standard "not an integer" error. Widened
Parser::integer_response to long long so large counters don't
overflow the RESP reply.

+++ include/parser.h
@@ -18,7 +18,7 @@ public:
     static std::string error_response(std::string_view msg);
     static std::string bulk_response(std::string_view value);
     static std::string null_response();
-    static std::string integer_response(int value);
+    static std::string integer_response(long long value);
 };
 
 } // namespace litekv
